move admin-user pages to admin/user - rename admin-user-access to batch-sync-user-access - rename admin-users to admin-user-table - rename admin-user to deprecated-admin-user
