+++ ui/src/app/pages/admin/user/admin-user-common.tsx
